Update MathJaxWrapper for RamButton.tsx
This is probably the most useless update that I contributed that I committed to a github project, oddly enough this is the first commit on a github project that I did (most github contributions from me are issues from other github projects)

Anyways, there is a difference between 1.58^{\log_2{(\text{RAM})}} and 1.58^{log_2{(ram)}}, in LaTeX when we use Logarithms we type out `\log` to indicate that it is a log function. As for \text{RAM}, it shows that this is a single variable, unlike the original `ram`, which could be interpreted as: `r \cdot a \cdot m`.

Therefore, I suggest updating the text to better reflect the formula.